examples/chat-python-flask/requirements.txt
Flask>=1.1.2
PyJWT>=1.7.1
sseclient>=0.0.26
gunicorn>=20.0.4
